+++ src/Sutil/TransitionFunctions.fs
@@ -22,7 +22,7 @@ let parseFloat (s:string, name) =
 let fade (initProps : TransitionProp list) (node : HTMLElement)  : CreateTransition = fun () ->
     let tr = applyProps initProps { Transition.Default with Delay = 0.0; Duration = 400.0; Ease = Easing.linear }
     {
-        tr with Css = (fun t _ -> sprintf "opacity: %f" (t* computedStyleOpacity node))
+        tr with Css = Some (fun t _ -> sprintf "opacity: %f" (t* computedStyleOpacity node))
     }
 
 let slide (props : TransitionProp list) (node : HTMLElement) =
@@ -41,7 +41,7 @@ let slide (props : TransitionProp list) (node : HTMLElement) =
 
     let set (name,value,units) = sprintf "%s: %s%s;" name value units
 
-    fun () -> { tr with Css = (fun t _ ->
+    fun () -> { tr with Css = Some(fun t _ ->
                             let result = ([
                                     ("overflow", "hidden", "")
                                     ("opacity",  (min (t * 20.0) 1.0) * opacity  |> string, "")
@@ -71,7 +71,7 @@ let draw (props : TransitionProp list) (node : SVGPathElement) =
 
     fun () -> { tr with
                     Duration = duration
-                    Css = fun t u -> sprintf "stroke-dasharray: %f %f" (t*len) (u*len) }
+                    Css = Some <| fun t u -> sprintf "stroke-dasharray: %f %f" (t*len) (u*len) }
 
 let fly (props : TransitionProp list) (node : Element) =
     let tr = applyProps props { Transition.Default with Delay = 0.0; Duration = 400.0; Ease = Easing.cubicOut; X = 0.0; Y = 0.0 }
@@ -84,7 +84,7 @@ let fly (props : TransitionProp list) (node : Element) =
     fun () -> {
         // Called when animation ready to run
         tr with
-            Css = (fun t u ->
+            Css = Some(fun t u ->
                 sprintf "transform: %s translate(%fpx, %fpx); opacity: %f;"
                         transform
                         ((1.0 - t) * tr.X)
@@ -127,7 +127,7 @@ let crossfade userProps =
             tr with
                 DurationFn = None
                 Duration = duration
-                Css = fun t u ->
+                Css = Some <| fun t u ->
                     sprintf """
                       opacity: %f;
                       transform-origin: top left;